Updated/Fixed Boss
Boss now properly incorperates art/mesh for segments. Some behavior adjustments made to increase functionality. Changed EntityBase health from Int to Float, and changed related scripts to update from int values to float values.
Also included a fix for Projectile script to fire projectiles at proper speeds instead of having a "slow batch" at the start of firing.

+++ Assets/_Game/Scripts/Game/EntityBase.cs
@@ -11,8 +11,8 @@ public class EntityBase : MonoBehaviour
     public UnityEvent Died;
 
     [Header("Settings")]
-    [SerializeField] protected int maxHealth = 1;
-    protected int _currentHealth = 0;
+    [SerializeField] protected float maxHealth = 1f;
+    protected float _currentHealth = 0f;
 
     // Start is called before the first frame update
     void Start()
@@ -24,7 +24,7 @@ public class EntityBase : MonoBehaviour
     /// 
     /// </summary>
     /// <param name="damage"> Value passed from Projectile/Source. Amount of Damage.</param>
-    public virtual void TakeDamage(int damage)
+    public virtual void TakeDamage(float damage)
     {
         _currentHealth -= damage;
